# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/customer": {
            "get": {
                "description": "Retrieves all customers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get all customers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Customer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing customer by customer number or ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Update customer",
                "parameters": [
                    {
                        "description": "Customer payload",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new customer to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Insert new customer",
                "parameters": [
                    {
                        "description": "Customer payload",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer-note": {
            "put": {
                "description": "Updates an existing customer note",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Update a customer note",
                "parameters": [
                    {
                        "description": "Customer Note Payload",
                        "name": "customerNote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerNote"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update result",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new customer note into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Create a new customer note",
                "parameters": [
                    {
                        "description": "Customer Note Payload",
                        "name": "customerNote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerNote"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Insert result",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer-note/get-all": {
            "get": {
                "description": "Retrieves all customer notes from the system",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Get all customer notes",
                "responses": {
                    "200": {
                        "description": "List of customer notes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.CustomerNote"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer-note/get-by-customer-number/{customer_number}": {
            "get": {
                "description": "Retrieves all notes associated with a given customer number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Get customer notes by customer number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Number",
                        "name": "customer_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer notes for the customer number",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.CustomerNote"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer-note/get-by-id/{id}": {
            "get": {
                "description": "Retrieves a customer note by its unique ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Get a customer note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer note",
                        "schema": {
                            "$ref": "#/definitions/domain.CustomerNote"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer-note/{id}": {
            "delete": {
                "description": "Deletes a customer note by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customer-note"
                ],
                "summary": "Delete a customer note by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Note ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete result",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/customer/{customer_number}": {
            "get": {
                "description": "Retrieves a customer by their customer number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get customer by number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Number",
                        "name": "customer_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Customer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a customer based on customer number",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Delete customer by number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Customer Number",
                        "name": "customer_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Customer": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "1995-06-12T00:00:00Z"
                },
                "created_at": {
                    "$ref": "#/definitions/types.NullTime"
                },
                "customer_number": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/types.NullTime"
                }
            }
        },
        "domain.CustomerNote": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/types.NullTime"
                },
                "customer_number": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statuscode": {
                    "type": "integer"
                }
            }
        },
        "types.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        }
    }
}